requirements.txt
Django==3.2.3
djangorestframework==3.12.4
requests==2.25.1
beautifulsoup4==4.9.3
python-dotenv==0.17.1
celery==5.0.5
gunicorn==20.1.0
psycopg2==2.8.6
redis==3.5.3
